+++ packages/canon/src/components/TextField/types.ts
@@ -15,8 +15,7 @@
  */
 
 import type { Breakpoint } from '../../types';
-import type { IconNames } from '../Icon';
-
+import type { ReactNode, MouseEventHandler } from 'react';
 /** @public */
 export interface TextFieldProps
   extends Omit<React.ComponentPropsWithoutRef<'input'>, 'size'> {
@@ -54,10 +53,10 @@ export interface TextFieldProps
   /**
    * An icon to render before the input
    */
-  icon?: IconNames;
+  icon?: ReactNode;
 
   /**
    * Handler to call when the clear button is pressed
    */
-  onClear?: React.MouseEventHandler<HTMLButtonElement>;
+  onClear?: MouseEventHandler<HTMLButtonElement>;
 }