+++ packages/frontend-app-api/src/wiring/createSpecializedApp.tsx
@@ -208,6 +208,7 @@ export function createSpecializedApp(options?: {
   extensionFactoryMiddleware?:
     | ExtensionFactoryMiddleware
     | ExtensionFactoryMiddleware[];
+  flags?: { allowUnknownExtensionConfig?: boolean };
 }): { apis: ApiHolder; tree: AppTree } {
   const config = options?.config ?? new ConfigReader({}, 'empty-config');
   const features = deduplicateFeatures(options?.features ?? []);
@@ -221,6 +222,7 @@ export function createSpecializedApp(options?: {
       ],
       parameters: readAppExtensionsConfig(config),
       forbidden: new Set(['root']),
+      allowUnknownExtensionConfig: options?.flags?.allowUnknownExtensionConfig,
     }),
   );