+++ packages/frontend-plugin-api/src/components/ExtensionBoundary.tsx
@@ -69,14 +69,14 @@ export function ExtensionBoundary(props: ExtensionBoundaryProps) {
     node.instance?.getData(coreExtensionData.routePath),
   );
 
-  const plugin = node.spec.source;
+  const plugin = node.spec.plugin;
   const Progress = useComponentRef(coreComponentRefs.progress);
   const fallback = useComponentRef(coreComponentRefs.errorBoundaryFallback);
 
   // Skipping "routeRef" attribute in the new system, the extension "id" should provide more insight
   const attributes = {
     extensionId: node.spec.id,
-    pluginId: node.spec.source?.id ?? 'app',
+    pluginId: node.spec.plugin?.id ?? 'app',
   };
 
   return (