+++ plugins/catalog-backend-module-gcp/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @backstage/plugin-catalog-backend-module-gcp
 
+## 0.3.8-next.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.3.1-next.2
+  - @backstage/catalog-model@1.7.3
+  - @backstage/config@1.3.2
+  - @backstage/plugin-catalog-node@1.17.0-next.2
+  - @backstage/plugin-kubernetes-common@0.9.5-next.0
+
 ## 0.3.8-next.1
 
 ### Patch Changes